Add files via upload
One of the monthly reports that our security team have to do monthly is called "archer report".  We had to work on different resources from excel, csv, html and picture files, do calculations in each of the files and gather the result into a report. It takes a lot of time and there can be inaccuracies. Therefore, I wrote the code to generate the report. It reduced the time to generate the report from 45 minutes last month to around 10 minutes this month(save 78%), avoid all human mistakes such as mistypes and create ease of use. With pyinstaller, it became an application so everyone in my team can use without installing python or its libraries. (the dist file is too heavy to be uploaded).
What I used: simple pandas, pytesseract to recognize words from images, and pyinstaller.
link for pytesseract https://pypi.org/project/pytesseract/
link for pyinstaller https://www.pyinstaller.org/