Archer report with user interface using tkinter. It takes in three entries, has a browse button to select folder and a "create" button to run the archer generator code.
Can be improved: add a pop up window when work is done,  add background, change the positions of buttons.
However, this year my boss told me we will switch to another kind of report so I will work on that and maybe not continue on this one. I may still reuse the tkinter part.